requirements.txt
torch>=2.2
torchvision>=0.17
timm>=1.0.7
scikit-learn>=1.4
pandas>=2.2
numpy>=1.26
opencv-python
pyyaml
matplotlib
